Rework the tests to run on pc86; lots of test fixes for the brk() test, which was nearly useless; lots of fixes to qemuppc and pc86 sbrk(), which was broken; change the pc86 console to echo output to the serial port (needed for running tests on qemu).

+++ tests/plat/testdriver.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-qemu=$1
+method=$1
 img=$2
 timeout=$3
 
@@ -13,10 +13,25 @@ trap "rm -f $result" EXIT
 pidfile=/tmp/$$.testdriver.pid
 trap "rm -f $pidfile" EXIT
 
-( $qemu -nographic -kernel $img 2>&1 & echo $! > $pidfile ) \
-    | tee $result \
-    | ( timeout $timeout grep -l -q @@FINISHED ; echo ) \
-    | ( read dummy && kill $(cat $pidfile) )
+case $method in
+    qemu-system-*)
+        if ! hash $method 2>/dev/null; then
+            echo "Warning: $method not installed, skipping test"
+            exit 0
+        fi
+
+        case $method in
+            qemu-system-i386) img="-drive file=$img,if=floppy,format=raw" ;;
+            qemu-system-ppc)  img="-kernel $img" ;;
+        esac
+
+        ( $method -nographic $img 2>&1 & echo $! > $pidfile ) \
+            | tee $result \
+            | ( timeout $timeout grep -l -q @@FINISHED ; echo ) \
+            | ( read dummy && kill $(cat $pidfile) )
+        
+        ;;
+esac
 
 ( grep -q @@FAIL $result || ! grep -q @@FINISHED $result ) && cat $result && exit 1
 exit 0
